added user registration, password hashing, JWT login, route protection, user & admin roles

+++ app/api/jobs.py
@@ -11,19 +11,24 @@ from app.schemas.job import JobCreate, JobResponse
 from app.execution.factory import ExecutorFactory
 from app.cache.job_cache import JobCache
 
+from app.auth.dependencies import get_current_user
+from app.models.user import Roles
+
 
 router = APIRouter(prefix = "/jobs", tags = ["Jobs"])
-job_repo = JobRepository(db)
+# job_repo = JobRepository(db)
 
 
 # depends on successful db session
 @router.post("/", response_model = JobResponse)
-def create_job(request: JobCreate, db: Session = Depends(get_db)):
+def create_job(request: JobCreate, db: Session = Depends(get_db), user = Depends(get_current_user)):
+	job_repo = JobRepository(db)
 	logger.info("API enters")
 	job = Job(name = request.name,
 				command = request.command,
 				job_type = request.job_type,
-				schedule_time = request.schedule_time)
+				schedule_time = request.schedule_time,
+				user_id = int(user["sub"]))
 	job = job_repo.create(job)
 	JobCache.put(job)
 	logger.info("after create API")
@@ -32,12 +37,14 @@ def create_job(request: JobCreate, db: Session = Depends(get_db)):
 
 # returning list of Jobs
 @router.get("/", response_model = list[JobResponse])
-def list_jobs(db: Session = Depends(get_db)):
-	return job_repo.get_all()
+def list_jobs(db: Session = Depends(get_db), user = Depends(get_current_user)):
+	job_repo = JobRepository(db)
+	return job_repo.get_all(user)
 
 
 @router.get("/{job_id}", response_model = JobResponse)
 def get_job(job_id: int, db: Session = Depends(get_db)):
+	job_repo = JobRepository(db)
 	job = job_repo.get(job_id)
 	if job is None:
 		raise HTTPException(status_code = 404, detail = f"Job {job_id} not found")
@@ -45,7 +52,12 @@ def get_job(job_id: int, db: Session = Depends(get_db)):
 
 
 @router.delete("/{job_id}")
-def delete_job(job_id: int, db: Session = Depends(get_db)):
+def delete_job(job_id: int, db: Session = Depends(get_db), user = Depends(get_current_user)):
+	job_repo = JobRepository(db)
+	job = job_repo.get(job_id)
+	if (job.user_id != int(user["sub"])) and (user["role"] != Roles.ADMIN):
+		raise HTTPException(403, "Forbidden")
+
 	deleted = job_repo.delete(job_id)
 	JobCache.invalidate(job_id)
 	if not deleted:
@@ -57,9 +69,16 @@ def delete_job(job_id: int, db: Session = Depends(get_db)):
 # temporary execute API before putting it in scheduler
 @router.post("/{job_id}/execute")
 def execute_job(job_id: int, db: Session = Depends(get_db)):
+	job_repo = JobRepository(db)
 	job = job_repo.get(job_id)
 	if not job:
 		raise HTTPException(404, f"Job {job_id} not found")
 	executor = ExecutorFactory.get(job.job_type)
 	result = executor.execute(job.command)
 	return result
+
+
+# WHO AM I - endpoint check
+@router.get("/me")
+def me(user = Depends(get_current_user)):
+	return user
